rework: xp is float now

+++ Un1t (second iteration)/Assets/Scripts/Player/PlayerConfig.cs
@@ -7,7 +7,7 @@ public class PlayerConfig : ScriptableObject
     [SerializeField] private float baseMaxHealth;
     [SerializeField] private int level;
     [Tooltip("How much XP player needs to get level")]
-    [SerializeField] private List<int> xpToNextLevel; // can't use IReadOnlyList because it doesn't support serialization
+    [SerializeField] private List<float> xpToNextLevel; // can't use IReadOnlyList because it doesn't support serialization
     [Tooltip("What part of NextLevelXP will be spent on healing")]
     [SerializeField, Range(0.1f, 1f)] private float baseHealCostCoefficient;
     [SerializeField] private float baseMovingSpeed;
@@ -16,7 +16,7 @@ public class PlayerConfig : ScriptableObject
     [SerializeField] private float baseDashCooldown;
     public float BaseMaxHealth => baseMaxHealth;
     public int Level => level;
-    public List<int> XPToNextLevel => xpToNextLevel;
+    public List<float> XPToNextLevel => xpToNextLevel;
     public float BaseHealCostCoefficient => baseHealCostCoefficient;
     public float BaseMovingSpeed => baseMovingSpeed;
     public float BaseDashSpeed => baseDashSpeed;
